+++ HDL/functions/values.js
@@ -1,0 +1,20 @@
+module.exports = {
+    list: {
+        ACValues: {
+            "tempType": {0: "C", 1: "F"},
+            "modes": {
+                0: "cooling",
+                1: "heating",
+                2: "fan",
+                3: "auto",
+                4: "dry"
+            },
+            "fanSpeeds": {
+                0: "auto",
+                1: "high",
+                2: "medium",
+                3: "low"
+            },
+        }
+    }
+}