+++ nodes/bus.js
@@ -140,14 +140,14 @@ module.exports = function(RED)
             //Send it!
             node.status({fill:"orange",shape:"dot",text:"Sending..."});
             network.send(node, msg, function(success, packet) {
-                if(success) {
+                if(success == true) {
                     node.status({fill:"green",shape:"dot",text:"Sent!"});
                     node.sendMessage(packet);
                 }
                 else {
                     node.status({fill:"red",shape:"dot",text:"Failed"});
                 }
-            });
+            }).catch((error)=>{});
         });
 
         //Add the node information to the msg object