+++ versionHistory.md
@@ -14,5 +14,5 @@
 - 1.6.2: Add the ability to override the subnet/device id for a device node
 - 1.7.0: Added blocking of duplicate messages
 - 1.8.0: 
-    - Updated panel control for better processing of the data
+    - Updated panel control for better processing of the data (disabled for the moment)
     - Added read temperature